+++ lib/src/Payment.dart
@@ -31,35 +31,40 @@ class Payment {
   String country;
   List<Link> links;
   String capturedDate;
+  int capturedAmount;
+  String voidedDate;
+  int voidedAmount;
 
-  Payment({
-    this.serviceTaxAmount,
-    this.installments,
-    this.interest,
-    this.capture,
-    this.authenticate,
-    this.recurrent,
-    this.creditCard,
-    this.tid,
-    this.proofOfSale,
-    this.authorizationCode,
-    this.softDescriptor,
-    this.provider,
-    this.fraudAnalysis,
-    this.isQrCode,
-    this.amount,
-    this.receivedDate,
-    this.status,
-    this.isSplitted,
-    this.returnMessage,
-    this.returnCode,
-    this.paymentId,
-    this.type,
-    this.currency,
-    this.country,
-    this.links,
-    this.capturedDate,
-  });
+  Payment(
+      {this.serviceTaxAmount,
+      this.installments,
+      this.interest,
+      this.capture,
+      this.authenticate,
+      this.recurrent,
+      this.creditCard,
+      this.tid,
+      this.proofOfSale,
+      this.authorizationCode,
+      this.softDescriptor,
+      this.provider,
+      this.fraudAnalysis,
+      this.isQrCode,
+      this.amount,
+      this.receivedDate,
+      this.status,
+      this.isSplitted,
+      this.returnMessage,
+      this.returnCode,
+      this.paymentId,
+      this.type,
+      this.currency,
+      this.country,
+      this.links,
+      this.capturedDate,
+      this.capturedAmount,
+      this.voidedDate,
+      this.voidedAmount});
 
   factory Payment.fromJson(Map<String, dynamic> json) =>
       _$PaymentFromJson(json);