+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [1.0.0] - 02 de Maio 2020.
+
+* Funções adicionadas:
+    * getReturn(paymentId) - Função para consulta
+    * enableCapture(paymentId) - Função para aprovar a compra
+    * enableVoid(paymentId) - Função para cancelar a compra
+* Variáveis adicionadas ao Payment:
+    * capturedAmount - Valor da aprovação
+    * voidedDate - Data do cancelamento
+    * voidedAmount - Valor cancelado
+
 ## [0.1.2] - 06 de Abril 2020.
 
 * Correções example